+++ src/yield/http/http_message.cpp
@@ -51,7 +51,7 @@ HTTPMessage<HTTPMessageType>::
 HTTPMessage(
   YO_NEW_REF Object* body,
   uint32_t connection_id,
-  float http_version
+  uint8_t http_version
 ) : body(body),
     connection_id(connection_id),
     header(*new Buffer(Buffer::getpagesize(), Buffer::getpagesize())),
@@ -66,7 +66,7 @@ HTTPMessage(
   uint32_t connection_id,
   uint16_t fields_offset,
   Buffer& header,
-  float http_version
+  uint8_t http_version
 ) : body(body),
     connection_id(connection_id),
     fields_offset(fields_offset),
